feat: add provider base classes
